Fix voltage

+++ src/vcu/inc/state_vcu.h
@@ -37,7 +37,7 @@ typedef struct {
 } MCReadings_T;
 
 typedef struct {
-  int16_t V_bus;
+  int32_t V_bus;
   int32_t power;
 } CSReadings_T;
